added UI module
